requirements.txt
numpy
pandas
python-dotenv
requests
scikit-learn
tqdm
